refactor: useCanvas hook 분리
