+++ src/pages/Archive/types/index.ts
@@ -1,0 +1,6 @@
+export interface State {
+  isDown: boolean;
+  moveX: number;
+  offsetX: number;
+  rotate: number;
+}
